refactor: extract game namespace socket handler
기본 '/' path 에서 사용하는 게임전용 소켓들 모두 추출.
하나씩 파일로 추출예정
